requirements.txt
alibabacloud_oss_v2==1.1.1
fastapi==0.115.12
openai==1.82.1
pandas==2.2.3
Pillow==11.2.1
PyYAML==6.0.2
torch==2.2.0
torchvision==0.17.0